support leveldb

+++ src/app/storage/storage.controller.ts
@@ -9,7 +9,10 @@ import {
 } from '@nestjs/common'
 import { Auth } from 'decorators/auth.decorator'
 import { SignatureGuard } from 'guards/sig.guard'
-import { ParseAddressPipe, ParseUserAuthPipe } from 'pipelines/address.pipeline'
+import {
+  ParseAddressPipe,
+  ParseSignerAuthPipe,
+} from 'pipelines/address.pipeline'
 import { StorageService } from './storage.service'
 
 @Controller('storage')
@@ -17,26 +20,36 @@ export class StorageController {
   constructor(private readonly service: StorageService) {}
 
   @Post()
-  async createStorage(@Query('multisig') multisigId: string) {
+  @UseGuards(SignatureGuard)
+  async createStorage(
+    @Auth(ParseSignerAuthPipe, ParseAddressPipe) multisigId: string,
+  ) {
     return this.service.createStorage('', { multisigId })
   }
 
   @Get()
   @UseGuards(SignatureGuard)
   async readStorage(
-    @Auth(ParseUserAuthPipe, ParseAddressPipe) userId: string,
+    @Auth(ParseSignerAuthPipe, ParseAddressPipe) userId: string,
     @Query('multisig') multisigId: string,
   ) {
     return this.service.readStorage(userId, { multisigId })
   }
 
   @Patch()
-  async updateStorage(@Query('multisig') multisigId: string) {
+  @UseGuards(SignatureGuard)
+  async updateStorage(
+    @Auth(ParseSignerAuthPipe, ParseAddressPipe) multisigId: string,
+  ) {
     return this.service.updateStorage('', { multisigId })
   }
 
   @Delete()
-  async deleteStorage(@Query('multisig') multisigId: string) {
-    return this.service.deleteStorage('', { multisigId })
+  @UseGuards(SignatureGuard)
+  async deleteStorage(
+    @Auth(ParseSignerAuthPipe, ParseAddressPipe) userId: string,
+    @Query('multisig', ParseAddressPipe) multisigId: string,
+  ) {
+    return this.service.deleteStorage(userId, { multisigId })
   }
 }
